Move event machinery into ppb.eventlib
