+++ ppb/engine.py
@@ -11,9 +11,7 @@ from typing import Type
 from typing import Union
 
 import ppb.events as events
-from ppb.events import StartScene
-from ppb.events import EventMixin
-from ppb.events import Quit
+from ppb.eventlib import EventMixin
 from ppb.systems import PygameEventPoller
 from ppb.systems import Renderer
 from ppb.systems import Updater
@@ -135,7 +133,7 @@ class GameEngine(EventMixin, LoggingMixin):
             for game_object in scene:
                 game_object.__event__(event, self.signal)
 
-    def on_start_scene(self, event: StartScene, signal: Callable[[Any], None]):
+    def on_start_scene(self, event: events.StartScene, signal: Callable[[Any], None]):
         """
         Start a new scene. The current scene pauses.
         """
@@ -159,7 +157,7 @@ class GameEngine(EventMixin, LoggingMixin):
         self.stop_scene()
         self.start_scene(event.new_scene, event.kwargs)
 
-    def on_quit(self, quit_event: Quit, signal: Callable[[Any], None]):
+    def on_quit(self, quit_event: events.Quit, signal: Callable[[Any], None]):
         self.running = False
 
     def pause_scene(self):
