+++ ppb/events.py
@@ -1,6 +1,7 @@
+"""
+A big pile of the events defined in PPB.
+"""
 from dataclasses import dataclass
-import logging
-import re
 from typing import Any
 from typing import Collection
 from typing import Dict
@@ -11,7 +12,6 @@ from typing import Union
 
 __all__ = (
     'StartScene',
-    'EventMixin',
     'PreRender',
     'Quit',
     'Render',
@@ -23,63 +23,6 @@ __all__ = (
     'StopScene',
     'Update',
 )
-
-boundaries_finder = re.compile('(.)([A-Z][a-z]+)')
-boundaries_finder_2 = re.compile('([a-z0-9])([A-Z])')
-
-def camel_to_snake(txt):
-    s1 = boundaries_finder.sub(r'\1_\2', txt)
-    return boundaries_finder_2.sub(r'\1_\2', s1).lower()
-
-
-class BadEventHandlerException(TypeError):
-
-    def __init__(self, instance, method, event):
-        object_type = type(instance)
-        event_type = type(event)
-        o_name = object_type.__name__
-        e_name = event_type.__name__
-        article = ['a', 'an'][int(e_name.lower()[0] in "aeiou")]
-
-        message = f"""
-{o_name}.{method}() signature incorrect, it should accept {article} {e_name} object and a signal function.
-
-{e_name} is a dataclass that represents an event. Its attributes 
-tell you about the event.
-
-The signal function is a function you can call that accepts an event instance
-as its only parameter. Call it to add an event to the queue. You don't have to
-use it, but it is a mandatory argument provided by ppb.
-
-It should look like this:
-
-def {method}({e_name.lower()}_event: {e_name}, signal_function):
-    (Your code goes here.)
-"""
-        super().__init__(message)
-
-
-class EventMixin:
-    def __event__(self, bag, fire_event):
-        elog = logging.getLogger('game.events')
-
-        name = camel_to_snake(type(bag).__name__)
-        meth_name = 'on_' + name
-        meth = getattr(self, meth_name, None)
-        if callable(meth):
-            try:
-                elog.debug(f"Calling handler {meth} for {name}")
-                meth(bag, fire_event)
-            except TypeError as ex:
-                from inspect import signature
-                sig = signature(meth)
-                try:
-                    sig.bind(bag, fire_event)
-                except TypeError:
-                    raise BadEventHandlerException(self, meth_name, bag) from ex
-                else:
-                    raise
-
 
 # Remember to define scene at the end so the pargs version of __init__() still works
